Update static_assurance_2.md
Alters "... its type system. Which ..." to "... its type system, which ..." in: "... the language itself, in its type system. Which implements a flavor ..." (line 200)

It may be better to have "... its type system. Which ..." instead become "... its type system. This type system ...", but I am uncertain if "Which" absolutely refers to the type system, so I've only suggested a minimal grammatical correction rather than a rewording, as per `CONTRIBUTING.md`